feat: add early finish to getMostAllowedBlockedScript script
imagine that a user is not using the deny list. There won't be any blocked items. But the getMostAllowedBlockedScript script will still try to find n many blocked items, going through all the items in the ZUNION result (which is potentiall huge).

With this change, we add an early finish. If checkAtMost=10 and itemCount=20, getMostAllowedBlocked method will go through 10 items and stop, instead of trying to find 30 (10 from each of success/ratelimited/denied)